+++ bigquery/src/main/scala/org/datatools/bigdatatypes/bigquery/BigQueryTypesInstance.scala
@@ -5,13 +5,15 @@ import org.datatools.bigdatatypes.conversions.SqlInstanceConversion
 import org.datatools.bigdatatypes.formats.Formats
 import org.datatools.bigdatatypes.types.basic.SqlType
 
-/** Type class to convert generic SqlTypes into BigQuery specific fields
+/** Type class to convert generic SqlTypes received as instance into BigQuery specific fields
+  * This uses [[BigQueryTypes]] to create BigQuery Fields
   *
   * @tparam A the type we want to obtain an schema from
   */
 trait BigQueryTypesInstance[A] {
 
-  /** @return a list of [[Field]]s that represents [[A]]
+  /** @param value an instance of [[A]]
+    * @return a list of [[Field]]s that represents [[A]]
     */
   def bigQueryFields(value: A): List[Field]
 }
@@ -22,14 +24,12 @@ object BigQueryTypesInstance {
     */
   def apply[A](implicit a: BigQueryTypesInstance[A]): BigQueryTypesInstance[A] = a
 
-  /** Factory constructor - allows easier construction of instances
-    */
-  def instance[A](f: A => List[Field]): BigQueryTypesInstance[A] = (value: A) => f(value)
-
   //TODO change it for the compressed syntax after having everything well documented
   /** Instance derivation via SqlTypeConversion. It uses `getSchema` from BigQueryTypes Type Class
     */
-  implicit def fieldsFromSqlInstanceConversion[A: SqlInstanceConversion](implicit f: Formats): BigQueryTypesInstance[A] =
+  implicit def fieldsFromSqlInstanceConversion[A: SqlInstanceConversion](implicit
+      f: Formats
+  ): BigQueryTypesInstance[A] =
     new BigQueryTypesInstance[A] {
 
       /** @return a list of [[Field]]s that represents [[A]]