+++ examples/src/test/scala/org/datatools/bigdatatypes/CrossModuleExamplesSpec.scala
@@ -1,6 +1,5 @@
 package org.datatools.bigdatatypes
 
-import com.google.cloud.bigquery.Field.Mode
 import org.apache.spark.sql.types.StructType
 import org.datatools.bigdatatypes.TestTypes._
 import org.datatools.bigdatatypes.bigquery.BigQueryTypesInstance
@@ -9,7 +8,6 @@ import org.datatools.bigdatatypes.formats.Formats.implicitDefaultFormats
 import org.datatools.bigdatatypes.spark.SparkTypes
 import org.datatools.bigdatatypes.types.basic.SqlType
 import org.datatools.bigdatatypes.spark.SqlTypeConversionSpark._
-import com.google.cloud.bigquery.{Field, StandardSQLTypeName}
 
 class CrossModuleExamplesSpec extends UnitSpec {