+++ spark/src/test/scala_2.13-/org/datatools/bigdatatypes/spark/SparkTypesSpec.scala
@@ -3,35 +3,16 @@ package org.datatools.bigdatatypes.spark
 import org.apache.spark.sql.types._
 import org.datatools.bigdatatypes.TestTypes._
 import org.datatools.bigdatatypes.{SparkTestTypes, UnitSpec}
-import org.datatools.bigdatatypes.formats.{DefaultFormats, Formats, SnakifyFormats}
+import org.datatools.bigdatatypes.formats.{DefaultFormats, Formats}
 import org.datatools.bigdatatypes.spark.SparkTypes._
 
+/** All conversions from Case Class to Spark Schemas
+  */
 class SparkTypesSpec extends UnitSpec {
 
-  behavior of "SparkTypesSpec"
   implicit val defaultFormats: Formats = DefaultFormats
 
-  case class Dummy(myInt: Int, myString: String)
-
-  val expectedFields: Seq[StructField] =
-    List(
-      StructField("myInt", IntegerType, nullable = false),
-      StructField("myString", StringType, nullable = false)
-    )
-  val expectedSchema: StructType = StructType(expectedFields)
-
-  "A Case Class instance" should "return Spark Fields" in {
-    val dummy = Dummy(1, "test")
-    dummy.sparkSchema shouldBe expectedSchema
-    dummy.sparkFields shouldBe expectedFields
-  }
-
-  "A Case Class type" should "return Spark Fields" in {
-    SparkTypes[Dummy].sparkSchema shouldBe expectedSchema
-    SparkTypes[Dummy].sparkFields shouldBe expectedFields
-    SparkSchemas.schema[Dummy] shouldBe expectedSchema
-    SparkSchemas.fields[Dummy] shouldBe expectedFields
-  }
+  behavior of "SparkTypesSpec for all TestTypes"
 
   "Basic types" should "create an Spark Schema" in {
     SparkSchemas.fields[BasicTypes] shouldBe SparkTestTypes.basicFields
@@ -62,85 +43,27 @@ class SparkTypesSpec extends UnitSpec {
     SparkSchemas.fields[ExtendedTypes] shouldBe SparkTestTypes.extendedTypes
   }
 
-  "Multiple Case Classes" should "create an Spark Schema with appended fields" in {
-    case class Append1(myAppend1: Int)
-    val fieldList: Seq[StructField] =
-      List(
-        StructField("myInt", IntegerType, nullable = false),
-        StructField("myString", StringType, nullable = false),
-        StructField("myAppend1", IntegerType, nullable = false)
-      )
-    SparkSchemas.fields[Dummy, Append1] shouldBe fieldList
-    SparkSchemas.schema[Dummy, Append1] shouldBe StructType(fieldList)
+  behavior of "SparkTypesSpec for syntactic sugars"
+
+  case class Dummy(myInt: Int, myString: String)
+
+  val expectedFields: Seq[StructField] =
+    List(
+      StructField("myInt", IntegerType, nullable = false),
+      StructField("myString", StringType, nullable = false)
+    )
+  val expectedSchema: StructType = StructType(expectedFields)
+
+  "A Case Class instance" should "return Spark Fields" in {
+    val dummy = Dummy(1, "test")
+    dummy.sparkSchema shouldBe expectedSchema
+    dummy.sparkFields shouldBe expectedFields
   }
 
-  "3 Case Classes" should "create an Spark Schema with appended fields" in {
-    case class Append1(myAppend1: Int)
-    case class Append2(myAppend2: String)
-    val fieldList: Seq[StructField] =
-      List(
-        StructField("myInt", IntegerType, nullable = false),
-        StructField("myString", StringType, nullable = false),
-        StructField("myAppend1", IntegerType, nullable = false),
-        StructField("myAppend2", StringType, nullable = false)
-      )
-    SparkSchemas.fields[Dummy, Append1, Append2] shouldBe fieldList
-    SparkSchemas.schema[Dummy, Append1, Append2] shouldBe StructType(fieldList)
-  }
-
-  "4 Case Classes" should "create an Spark Schema with appended fields" in {
-    case class Append1(myAppend1: Int)
-    case class Append2(myAppend2: String)
-    case class Append3(myAppend3: String)
-    val fieldList: Seq[StructField] =
-      List(
-        StructField("myInt", IntegerType, nullable = false),
-        StructField("myString", StringType, nullable = false),
-        StructField("myAppend1", IntegerType, nullable = false),
-        StructField("myAppend2", StringType, nullable = false),
-        StructField("myAppend3", StringType, nullable = false)
-      )
-    SparkSchemas.fields[Dummy, Append1, Append2, Append3] shouldBe fieldList
-    SparkSchemas.schema[Dummy, Append1, Append2, Append3] shouldBe StructType(fieldList)
-  }
-
-  "5 Case Classes" should "create an Spark Schema with appended fields" in {
-    case class Append1(myAppend1: Int)
-    case class Append2(myAppend2: String)
-    case class Append3(myAppend3: String)
-    case class Append4(myAppend4: String)
-    val fieldList: Seq[StructField] =
-      List(
-        StructField("myInt", IntegerType, nullable = false),
-        StructField("myString", StringType, nullable = false),
-        StructField("myAppend1", IntegerType, nullable = false),
-        StructField("myAppend2", StringType, nullable = false),
-        StructField("myAppend3", StringType, nullable = false),
-        StructField("myAppend4", StringType, nullable = false)
-      )
-    SparkSchemas.fields[Dummy, Append1, Append2, Append3, Append4] shouldBe fieldList
-    SparkSchemas.schema[Dummy, Append1, Append2, Append3, Append4] shouldBe StructType(fieldList)
-  }
-
-}
-
-class SparkTypesSnakifiedSpec extends UnitSpec {
-
-  behavior of "SparkTypesSnakifiedSpec"
-
-  "Field Transformations" should "be applied to field names" in {
-    case class Dummy(myInt: Int, myString: String)
-
-    val expectedFields: Seq[StructField] =
-      List(
-        StructField("my_int", IntegerType, nullable = false),
-        StructField("my_string", StringType, nullable = false)
-      )
-    val expectedSchema: StructType = StructType(expectedFields)
-
-    implicit val formats: Formats = SnakifyFormats
-    SparkSchemas.fields[Dummy] shouldBe expectedSchema
-    println(SparkSchemas.fields[Dummy])
+  "A Case Class type" should "return Spark Fields" in {
+    SparkTypes[Dummy].sparkSchema shouldBe expectedSchema
+    SparkTypes[Dummy].sparkFields shouldBe expectedFields
+    SparkSchemas.schema[Dummy] shouldBe expectedSchema
     SparkSchemas.fields[Dummy] shouldBe expectedFields
   }
 }